WebApp/webapp_requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
pydantic==2.8.2
